200 rpm torque command modification
Commented out cancellation of regen for rpm < 200.  This is actually
quite dangerous.
If brake input is used for regen, you get strong regenerative braking
for a given pedal input.  But as rpm decreases below 200, the regen is
suddenly released and the car surges forward suddenly - just as you
slow down to stop before a car in front of you at a stoplight.